requirements.txt
asgiref==3.8.1
Django==5.0.3
django-ckeditor-5==0.2.12
pillow==10.2.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0